un-component EventList so the diff is cleaner
