+++ src/components/events/index.tsx
@@ -4,5 +4,4 @@ export { default as EventCard } from './EventCard';
 export { default as EventCarousel } from './EventCarousel';
 export { default as EventDisplay } from './EventDisplay';
 export { DEFAULT_FILTER_STATE, default as EventFilter } from './EventFilter';
-export { default as EventList } from './EventList';
 export { default as EventModal } from './EventModal';
